+++ internal/alertmanager/tls.go
@@ -3,16 +3,13 @@ package alertmanager
 import (
 	"crypto/tls"
 	"crypto/x509"
+	"log/slog"
 	"net/http"
 	"os"
-
-	"github.com/rs/zerolog/log"
 )
 
 func configureTLSRootCAs(tlsConfig *tls.Config, caPath string) error {
-	log.Debug().
-		Str("path", caPath).
-		Msg("Loading TLS CA cert")
+	slog.Debug("Loading TLS CA cert", slog.String("path", caPath))
 	caCert, err := os.ReadFile(caPath)
 	if err != nil {
 		return err
@@ -24,13 +21,10 @@ func configureTLSRootCAs(tlsConfig *tls.Config, caPath string) error {
 }
 
 func configureTLSClientCert(tlsConfig *tls.Config, certPath, keyPath string) error {
-	log.Debug().
-		Str("cert", certPath).
-		Str("key", keyPath).
-		Msg("Loading TLS cert and key")
+	slog.Debug("Loading TLS cert and key", slog.String("cert", certPath), slog.String("key", keyPath))
 	cert, err := tls.LoadX509KeyPair(certPath, keyPath)
 	if err != nil {
-		log.Debug().Err(err).Msg("Failed to load TLS cert and key")
+		slog.Debug("Failed to load TLS cert and key", slog.Any("error", err))
 		return err
 	}
 	tlsConfig.Certificates = []tls.Certificate{cert}