+++ scripts/test-main.sh
@@ -23,7 +23,7 @@ ALERTMANAGER_URI=http://localhost \
 ALERTMANAGER_INTERVAL=1s \
 LISTEN_ADDRESS=127.0.0.1 \
 LISTEN_PORT=0 \
-LOG_LEVEL=fatal \
+LOG_LEVEL=error \
 LOG_CONFIG=false \
   ./karma.test \
   -test.run '^TestRunMain$' \